Added walls & deathmatch spawner
I added walls, ceilings and tried to seal most of them all. The faces for everything are also sorted out and a video was made for class to show off the blockout. I also added spawns for players theres pretty much three for each corner around the three rooms and a few on the top of the center point.